Update org-mode synchronization to not sync "beeminder-value" for habits
The previous behaviour was for all properties to be synchronized when an
org-mode headline is closed. This also updated the value to reflect the current
goal value on `beeminder.com`.

This could be an issue for org-mode habits where you are trying to add a
specific datapoint whenever the habit is completed, as the value would change
every time the habit is completed.

This change adds the property `beeminder-excluded-habit-sync-properties`, which
can be used to exclude certain habit properties from synchronization.